refactor controller and model
